backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
httpx
python-multipart
pillow